+++ pkg/ingester/metrics_test.go
@@ -40,6 +40,54 @@ func TestTSDBMetrics(t *testing.T) {
 			# 4*(12345 + 85787 + 999)
 			cortex_ingester_shipper_upload_failures_total 396524
 
+			# HELP cortex_ingester_tsdb_compactions_total Total number of TSDB compactions that were executed.
+			# TYPE cortex_ingester_tsdb_compactions_total counter
+			cortex_ingester_tsdb_compactions_total 693917
+
+			# HELP cortex_ingester_tsdb_compaction_duration_seconds Duration of TSDB compaction runs.
+			# TYPE cortex_ingester_tsdb_compaction_duration_seconds histogram
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="1"} 0
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="2"} 0
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="4"} 0
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="8"} 0
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="16"} 3
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="32"} 3
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="64"} 3
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="128"} 3
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="256"} 3
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="512"} 3
+			cortex_ingester_tsdb_compaction_duration_seconds_bucket{le="+Inf"} 3
+			cortex_ingester_tsdb_compaction_duration_seconds_sum 27
+			cortex_ingester_tsdb_compaction_duration_seconds_count 3
+
+			# HELP cortex_ingester_tsdb_wal_fsync_duration_seconds Duration of TSDB WAL fsync.
+			# TYPE cortex_ingester_tsdb_wal_fsync_duration_seconds summary
+			cortex_ingester_tsdb_wal_fsync_duration_seconds{quantile="0.5"} 30
+			cortex_ingester_tsdb_wal_fsync_duration_seconds{quantile="0.9"} 30
+			cortex_ingester_tsdb_wal_fsync_duration_seconds{quantile="0.99"} 30
+			cortex_ingester_tsdb_wal_fsync_duration_seconds_sum 30
+			cortex_ingester_tsdb_wal_fsync_duration_seconds_count 3
+
+			# HELP cortex_ingester_tsdb_wal_page_flushes_total Total number of TSDB WAL page flushes.
+			# TYPE cortex_ingester_tsdb_wal_page_flushes_total counter
+			cortex_ingester_tsdb_wal_page_flushes_total 1090441
+
+			# HELP cortex_ingester_tsdb_wal_completed_pages_total Total number of TSDB WAL completed pages.
+			# TYPE cortex_ingester_tsdb_wal_completed_pages_total counter
+			cortex_ingester_tsdb_wal_completed_pages_total 1189572
+
+			# HELP cortex_ingester_tsdb_wal_truncations_failed_total Total number of TSDB WAL truncations that failed.
+			# TYPE cortex_ingester_tsdb_wal_truncations_failed_total counter
+			cortex_ingester_tsdb_wal_truncations_failed_total 1288703
+
+			# HELP cortex_ingester_tsdb_wal_truncations_total Total number of TSDB  WAL truncations attempted.
+			# TYPE cortex_ingester_tsdb_wal_truncations_total counter
+			cortex_ingester_tsdb_wal_truncations_total 1387834
+
+			# HELP cortex_ingester_tsdb_wal_writes_failed_total Total number of TSDB WAL writes that failed.
+			# TYPE cortex_ingester_tsdb_wal_writes_failed_total counter
+			cortex_ingester_tsdb_wal_writes_failed_total 1486965
+
 			# HELP cortex_ingester_memory_series_created_total The total number of series that were created per user.
 			# TYPE cortex_ingester_memory_series_created_total counter
 			# 5 * (12345, 85787 and 999 respectively)
@@ -60,7 +108,7 @@ func TestTSDBMetrics(t *testing.T) {
 func populateTSDBMetrics(base float64) *prometheus.Registry {
 	r := prometheus.NewRegistry()
 
-	// shipper
+	// Thanos shipper.
 	dirSyncs := promauto.With(r).NewCounter(prometheus.CounterOpts{
 		Name: "thanos_shipper_dir_syncs_total",
 		Help: "Total number of dir syncs",
@@ -95,6 +143,56 @@ func populateTSDBMetrics(base float64) *prometheus.Registry {
 		Name: "prometheus_tsdb_head_series_removed_total",
 	})
 	seriesRemoved.Add(6 * base)
+
+	ran := promauto.With(r).NewCounter(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_compactions_total",
+		Help: "Total number of compactions that were executed for the partition.",
+	})
+	ran.Add(7 * base)
+
+	duration := promauto.With(r).NewHistogram(prometheus.HistogramOpts{
+		Name:    "prometheus_tsdb_compaction_duration_seconds",
+		Help:    "Duration of compaction runs",
+		Buckets: prometheus.ExponentialBuckets(1, 2, 10),
+	})
+	duration.Observe(9)
+
+	fsyncDuration := promauto.With(r).NewSummary(prometheus.SummaryOpts{
+		Name:       "prometheus_tsdb_wal_fsync_duration_seconds",
+		Help:       "Duration of WAL fsync.",
+		Objectives: map[float64]float64{0.5: 0.05, 0.9: 0.01, 0.99: 0.001},
+	})
+	fsyncDuration.Observe(10)
+
+	pageFlushes := promauto.With(r).NewCounter(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_wal_page_flushes_total",
+		Help: "Total number of page flushes.",
+	})
+	pageFlushes.Add(11 * base)
+
+	pageCompletions := promauto.With(r).NewCounter(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_wal_completed_pages_total",
+		Help: "Total number of completed pages.",
+	})
+	pageCompletions.Add(12 * base)
+
+	truncateFail := promauto.With(r).NewCounter(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_wal_truncations_failed_total",
+		Help: "Total number of WAL truncations that failed.",
+	})
+	truncateFail.Add(13 * base)
+
+	truncateTotal := promauto.With(r).NewCounter(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_wal_truncations_total",
+		Help: "Total number of WAL truncations attempted.",
+	})
+	truncateTotal.Add(14 * base)
+
+	writesFailed := promauto.With(r).NewCounter(prometheus.CounterOpts{
+		Name: "prometheus_tsdb_wal_writes_failed_total",
+		Help: "Total number of WAL writes that failed.",
+	})
+	writesFailed.Add(15 * base)
 
 	return r
 }